This is a combination of 2 commits.
This is a combination of 2 commits.

This is a combination of 2 commits.

new comment

edit file1

This is a combination of 3 commits.
file2 reduced to 3 lines

add f2

f5 add

This is a combination of 2 commits.

add c1

This is a combination of 4 commits.
add c2

add c3

add c4

add c5